+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add flag -u/--url to scan subcommand
 - Add flag -s/--silent to scan subcommand
+- Add flag -o/--output to scan subcommand
 - Add GitHub repository url support for scan subcommand
 - Support npm yarn.lock files
 - Support npm pnpm-lock.yaml files